requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9       # Alembic migrations (sync)
alembic==1.13.2

# ── Pydantic / Settings ───────────────────────────────────────────────────────
pydantic==2.8.0
pydantic-settings==2.3.4
pydantic[email]==2.8.0
email-validator==2.2.0

# ── Authentication ────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4

# ── LangChain + LLM Integrations ─────────────────────────────────────────────
langchain==0.2.6
langchain-ollama==0.1.3          # Ollama embeddings (local, free)
langchain-google-genai==1.0.7    # Google Gemini Flash LLM
langchain-community==0.2.6
tenacity>=8.2.0                  # Exponential backoff for API rate limits

# ── ChromaDB ─────────────────────────────────────────────────────────────────
chromadb==0.5.3

# ── Document Loading ──────────────────────────────────────────────────────────
pypdf==4.2.0                 # PDF loading
python-docx==1.1.2           # DOCX loading
docx2txt==0.8                # Alternative DOCX text extraction
aiofiles==23.2.1             # Async file I/O

# ── Excel Export ─────────────────────────────────────────────────────────────
openpyxl==3.1.5

# ── Logging ───────────────────────────────────────────────────────────────────
structlog==24.2.0

# ── File Handling ─────────────────────────────────────────────────────────────
python-multipart==0.0.9     # FastAPI file upload support

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.27.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0                # Already listed — for TestClient
